requirements.txt
langchain
pinecone-client
python-dotenv
pypdf
openai
tiktoken